requirements-test.txt
# Test script dependencies
requests>=2.31.0
pillow>=10.0.0

